Update Fruit database

+++ src/models/fruit.cr
@@ -1,6 +1,6 @@
 struct Fruit
-  def initialize(@id : Int64?, @no : String?, @description : String?)
+  def initialize(@id : Int32?, @no : String?, @description : String?)
   end
 
-  JSON.mapping(id: Int64?, no: String?, description: String?)
+  JSON.mapping(id: Int32?, no: String?, description: String?)
 end
